+++ lib/output/emoji.go
@@ -2,10 +2,14 @@ package output
 
 // Standard emoji for use in output.
 const (
-	EmojiFailure   = "❌"
-	EmojiWarning   = "❗️"
-	EmojiSuccess   = "✅"
-	EmojiInfo      = "ℹ️"
-	EmojiLightbulb = "💡"
-	EmojiAsterisk  = "✱"
+	EmojiFailure          = "❌"
+	EmojiWarning          = "❗️"
+	EmojiSuccess          = "✅"
+	EmojiInfo             = "ℹ️"
+	EmojiLightbulb        = "💡"
+	EmojiAsterisk         = "✱"
+	EmojiWarningSign      = "⚠️"
+	EmojiFingerPointRight = "👉"
+	EmojiOk               = "👌"
+	EmojiQuestionMark     = "❔"
 )